Added correct variables to .gitignore and rewrote the utility functions
